// src/interfaces/response.ts
// Generated by https://quicktype.io

export interface Response {
    futbolistas: Futbolista[];
    first:       boolean;
    last:        boolean;
    totalPages:  number;
}

export interface Futbolista {
    id:              number;
    nombres:         string;
    apellidos:       string;
    fechaNacimiento: string;
    caracteristicas: string;
    posicion:        string;
}

export type FutbolistaKey = keyof Futbolista